fastapi_app/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = ""
authors = ["Antonio Zea Jr <zeajantonio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
python-multipart = "^0.0.9"
structlog = "^24.1.0"
pydantic-settings = "^2.2.1"
tensorflow = "^2.16.1"
numpy = "^1.26.4"
scikit-learn = "^1.4.1.post1"
pillow = "^10.2.0"
sqlalchemy = "^2.0.28"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
prospector = "^1.10.3"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"